+++ src/main/java/com/telepathicgrunt/the_bumblezone/enchantments/PotentPoisonEnchantment.java
@@ -5,6 +5,7 @@ import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.MobType;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TridentItem;
 import net.minecraft.world.item.enchantment.Enchantment;
@@ -43,7 +44,7 @@ public class PotentPoisonEnchantment extends Enchantment {
 
     @Override
     public void doPostAttack(LivingEntity attacker, Entity victim, int level) {
-        if(victim instanceof LivingEntity livingEntity) {
+        if(victim instanceof LivingEntity livingEntity && livingEntity.getMobType() != MobType.UNDEAD) {
             livingEntity.addEffect(new MobEffectInstance(
                     MobEffects.POISON,
                     200,