+++ src/main/java/com/telepathicgrunt/the_bumblezone/modinit/BzBlockEntities.java
@@ -1,7 +1,7 @@
 package com.telepathicgrunt.the_bumblezone.modinit;
 
 import com.telepathicgrunt.the_bumblezone.Bumblezone;
-import com.telepathicgrunt.the_bumblezone.blocks.blockentities.GlazedCocoonBlockEntity;
+import com.telepathicgrunt.the_bumblezone.blocks.blockentities.HoneyCocoonBlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
@@ -12,5 +12,5 @@ public class BzBlockEntities {
     public static final DeferredRegister<BlockEntityType<?>> BLOCK_ENTITIES = DeferredRegister.create(ForgeRegistries.BLOCK_ENTITIES, Bumblezone.MODID);
 
     //Blocks
-    public static final RegistryObject<BlockEntityType<?>> GLAZED_COCOON_BE = BLOCK_ENTITIES.register("glazed_cocoon", () -> BlockEntityType.Builder.of(GlazedCocoonBlockEntity::new, BzBlocks.GLAZED_COCOON.get()).build(null));
+    public static final RegistryObject<BlockEntityType<?>> HONEY_COCOON_BE = BLOCK_ENTITIES.register("honey_cocoon", () -> BlockEntityType.Builder.of(HoneyCocoonBlockEntity::new, BzBlocks.HONEY_COCOON.get()).build(null));
 }