+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 ### **(V.4.3.0 Changes) (1.18.1 Minecraft)**
 
-- glazed cocoon (test breaking later)
+- honey cocoon (test breaking later)
 - stinger spear (test durability later) 
 - Paralyzed
-- potent poison
-- neurotoxin
+- potent poison (dont forget description)
+- neurotoxin (dont forget description)
 
 ##### Blocks:
 Fixed some honey filled blocks not consuming Glass Bottles when collecting Honey Bottles from the blocks.